requirements.txt
setuptools>=65.5.0
wheel>=0.38.0
streamlit>=1.20.0
pandas>=1.5.0
numpy>=1.20.0
scikit-learn>=1.0.0
torch>=1.13.0
matplotlib>=3.5.0
plotly>=5.10.0
pillow>=9.0.0 